:sparkles: add DeploymentTypeLogo component to display deployment type and version information in the sidebar
