+++ PuppyFlow/app/components/sidebar/Header.tsx
@@ -22,7 +22,6 @@ function Header({ setFlowFullScreen }: HeaderProps) {
   const { userName } = useWorkspaces()
   const settingsDialogRef = useRef<HTMLDialogElement>(null)
   const [activeTab, setActiveTab] = useState<'settings' | 'models' | 'billing' | 'servers'>('settings')
-  const [showVersionPopup, setShowVersionPopup] = useState(false)
 
   const handleCloseDialog = () => {
     settingsDialogRef.current?.close()
@@ -30,10 +29,6 @@ function Header({ setFlowFullScreen }: HeaderProps) {
 
   const handleSettingsClick = () => {
     settingsDialogRef.current?.showModal()
-  }
-
-  const handleDeploymentTypeClick = () => {
-    setShowVersionPopup(!showVersionPopup)
   }
 
   return (
@@ -48,33 +43,12 @@ function Header({ setFlowFullScreen }: HeaderProps) {
         />
 
         <div className="flex items-center gap-2">
-          <div className="relative">
-            <span 
-              className="flex items-center gap-1.5 px-2 py-1 text-[11px] text-[#5D6065] rounded-md font-medium group hover:bg-[#313131] transition-colors duration-200 cursor-pointer select-none"
-              onClick={handleDeploymentTypeClick}
-            >
-              <span className="group-hover:text-white transition-colors duration-200 select-none">
-                {process.env.NEXT_PUBLIC_DEPLOYMENT_TYPE}
-              </span>
-            </span>
-            
-            {showVersionPopup && (
-              <div className="absolute top-full left-0 mt-3 bg-[#2A2A2A] border border-[#404040] rounded-md p-2 shadow-lg z-10 min-w-[96px] text-[11px] select-none before:content-[''] before:absolute before:top-[-6px] before:left-3 before:w-3 before:h-3 before:bg-[#2A2A2A] before:border-t before:border-l before:border-[#404040] before:rotate-45 before:transform">
-                <div className="flex flex-col space-y-1 text-[#AAAAAA]">
-                  <div>Type: <span className="text-white">{process.env.NEXT_PUBLIC_DEPLOYMENT_TYPE}</span></div>
-                  <div>Version: <span className="text-white">{process.env.NEXT_PUBLIC_FRONTEND_VERSION}</span></div>
-                </div>
-              </div>
-            )}
-          </div>
-
           <button className='w-[32px] h-[32px] flex items-center justify-center group' onClick={handleSettingsClick}>
             <svg width="32" height="32" viewBox="0 0 32 32" fill="none" xmlns="http://www.w3.org/2000/svg" className="group-hover:bg-[#313131] rounded-md">
               <path fillRule="evenodd" clipRule="evenodd" d="M16 13.3333C14.5272 13.3333 13.3333 14.5272 13.3333 16C13.3333 17.4728 14.5272 18.6667 16 18.6667C17.4728 18.6667 18.6667 17.4728 18.6667 16C18.6667 14.5272 17.4728 13.3333 16 13.3333ZM14.6667 16C14.6667 15.2636 15.2636 14.6667 16 14.6667C16.7364 14.6667 17.3333 15.2636 17.3333 16C17.3333 16.7364 16.7364 17.3333 16 17.3333C15.2636 17.3333 14.6667 16.7364 14.6667 16Z" fill="#5D6065" className="group-hover:fill-[#FFFFFF]" />
               <path fillRule="evenodd" clipRule="evenodd" d="M15.7278 8C14.7325 8 13.8887 8.73186 13.7479 9.71716L13.6579 10.3474L13.1479 9.9649C12.3517 9.36772 11.2375 9.4469 10.5337 10.1507L10.1488 10.5356C9.445 11.2394 9.36582 12.3536 9.963 13.1498L10.3444 13.6583L9.71716 13.7479C8.73186 13.8887 8 14.7325 8 15.7278V16.2722C8 17.2675 8.73186 18.1113 9.71716 18.2521L10.3457 18.3419L9.9645 18.8502C9.36732 19.6464 9.4465 20.7606 10.1503 21.4644L10.5352 21.8493C11.239 22.5531 12.3532 22.6323 13.1494 22.0351L13.658 21.6536L13.7479 22.2828C13.8887 23.2681 14.7325 24 15.7278 24H16.2722C17.2675 24 18.1113 23.2681 18.2521 22.2828L18.3417 21.6552L18.8483 22.0351C19.6445 22.6323 20.7587 22.5531 21.4625 21.8493L21.8474 21.4644C22.5512 20.7606 22.6304 19.6464 22.0332 18.8502L21.6522 18.3422L22.2828 18.2521C23.2681 18.1113 24 17.2675 24 16.2722V15.7278C24 14.7325 23.2681 13.8887 22.2828 13.7479L21.6535 13.658L22.0347 13.1498C22.6319 12.3535 22.5527 11.2394 21.8489 10.5356L21.464 10.1507C20.7602 9.44688 19.646 9.36769 18.8498 9.96487L18.3419 10.3458L18.2521 9.71716C18.1113 8.73186 17.2675 8 16.2722 8H15.7278ZM15.0679 9.90572C15.1148 9.57729 15.3961 9.33333 15.7278 9.33333H16.2722C16.6039 9.33333 16.8852 9.57729 16.9321 9.90572L17.0966 11.057C17.1319 11.3037 17.3018 11.5101 17.5371 11.5922C17.7057 11.651 17.8699 11.7192 18.0291 11.7962C18.2535 11.9048 18.5199 11.879 18.7193 11.7294L19.6498 11.0315C19.9152 10.8325 20.2866 10.8589 20.5212 11.0935L20.9061 11.4784C21.1407 11.713 21.1671 12.0844 20.968 12.3498L20.2703 13.28C20.1207 13.4796 20.0949 13.7459 20.2036 13.9704C20.2807 14.1298 20.349 14.2942 20.4078 14.4629C20.4899 14.6982 20.6963 14.8681 20.943 14.9034L22.0943 15.0679C22.4227 15.1148 22.6667 15.3961 22.6667 15.7278V16.2722C22.6667 16.6039 22.4227 16.8852 22.0943 16.9321L20.943 17.0966C20.6963 17.1319 20.4899 17.3018 20.4078 17.5371C20.3488 17.7062 20.2804 17.8709 20.2031 18.0306C20.0944 18.2552 20.1201 18.5216 20.2698 18.7212L20.9665 19.6502C21.1656 19.9156 21.1392 20.287 20.9046 20.5216L20.5197 20.9065C20.2851 21.1411 19.9137 21.1675 19.6483 20.9684L18.7185 20.2711C18.519 20.1215 18.2528 20.0957 18.0283 20.2042C17.8693 20.281 17.7054 20.3491 17.5371 20.4078C17.3018 20.4899 17.1319 20.6963 17.0966 20.943L16.9321 22.0943C16.8852 22.4227 16.6039 22.6667 16.2722 22.6667H15.7278C15.3961 22.6667 15.1148 22.4227 15.0679 22.0943L14.9034 20.943C14.8681 20.6963 14.6982 20.4899 14.4629 20.4078C14.2942 20.349 14.1298 20.2807 13.9705 20.2036C13.746 20.095 13.4796 20.1207 13.2801 20.2704L12.3494 20.9684C12.084 21.1675 11.7126 21.1411 11.478 20.9065L11.0931 20.5216C10.8585 20.287 10.8321 19.9156 11.0312 19.6502L11.7293 18.7193C11.8789 18.5198 11.9047 18.2535 11.7961 18.029C11.7192 17.8698 11.651 17.7056 11.5922 17.5371C11.5101 17.3018 11.3037 17.1319 11.057 17.0966L9.90572 16.9321C9.57729 16.8852 9.33333 16.6039 9.33333 16.2722V15.7278C9.33333 15.3961 9.57729 15.1148 9.90572 15.0679L11.057 14.9034C11.3037 14.8681 11.5101 14.6982 11.5922 14.4629C11.6508 14.2947 11.7189 14.1309 11.7956 13.972C11.9041 13.7476 11.8783 13.4814 11.7288 13.2819L11.0297 12.3498C10.8306 12.0844 10.857 11.713 11.0916 11.4784L11.4765 11.0935C11.7111 10.8589 12.0825 10.8325 12.3479 11.0316L13.2792 11.7301C13.4788 11.8797 13.7452 11.9055 13.9697 11.7968C14.1293 11.7195 14.2939 11.6511 14.4629 11.5922C14.6982 11.5101 14.8681 11.3037 14.9034 11.057L15.0679 9.90572Z" fill="#5D6065" className="group-hover:fill-[#FFFFFF]" />
             </svg>
           </button>
-
 
           {/* Right section */}
           <button className='w-[32px] h-[32px] flex items-center justify-center group' onClick={() => setFlowFullScreen(false)}>
